refactor(EatInOrder) : EatOrder 리팩토링 준비 완료.
