--- src/main/java/kitchenpos/eatinorders/domain/OrderRepository.java
@@ -1,18 +1,0 @@
-package kitchenpos.eatinorders.domain;
-
-import kitchenpos.eatinorders.todo.domain.ordertables.OrderTable;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface OrderRepository {
-    Order save(Order order);
-
-    Optional<Order> findById(UUID id);
-
-    List<Order> findAll();
-
-    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
-}
-
