--- src/main/java/kitchenpos/eatinorders/infra/JpaOrderRepository.java
@@ -1,10 +1,0 @@
-package kitchenpos.eatinorders.infra;
-
-import kitchenpos.eatinorders.domain.Order;
-import kitchenpos.eatinorders.domain.OrderRepository;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
-}
