+++ src/main/java/kitchenpos/eatinorders/infra/OrderClientImpl.java
@@ -1,20 +1,20 @@
 package kitchenpos.eatinorders.infra;
 
-import kitchenpos.eatinorders.domain.OrderRepository;
-import kitchenpos.eatinorders.domain.OrderStatus;
+import kitchenpos.eatinorders.todo.domain.orders.EatInOrderRepository;
+import kitchenpos.eatinorders.todo.domain.orders.EatInOrderStatus;
 import kitchenpos.eatinorders.todo.domain.ordertables.OrderClient;
 import kitchenpos.eatinorders.todo.domain.ordertables.OrderTable;
 import org.springframework.stereotype.Component;
 
 @Component
 public class OrderClientImpl implements OrderClient {
-    private final OrderRepository orderRepository;
+    private final EatInOrderRepository orderRepository;
 
-    public OrderClientImpl(OrderRepository orderRepository) {
+    public OrderClientImpl(EatInOrderRepository orderRepository) {
         this.orderRepository = orderRepository;
     }
 
     public boolean containsInvalidOrderForClearOrderTable(OrderTable orderTable) {
-        return orderRepository.existsByOrderTableAndStatusNot(orderTable, OrderStatus.COMPLETED);
+        return orderRepository.existsByOrderTableAndStatusNot(orderTable, EatInOrderStatus.COMPLETED);
     }
 }
