+++ src/main/java/kitchenpos/eatinorders/todo/domain/orders/EatInOrder.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.todo.domain.orders;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -14,54 +14,46 @@ import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
 import kitchenpos.eatinorders.todo.domain.ordertables.OrderTable;
 import kitchenpos.support.domain.OrderLineItem;
-import kitchenpos.support.domain.OrderType;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
-@Table(name = "orders")
+@Table(name = "eat_in_orders")
 @Entity
-public class Order {
+public class EatInOrder {
     @Column(name = "id", columnDefinition = "binary(16)")
     @Id
     private UUID id;
 
-    @Column(name = "type", nullable = false, columnDefinition = "varchar(255)")
-    @Enumerated(EnumType.STRING)
-    private OrderType type;
-
     @Column(name = "status", nullable = false, columnDefinition = "varchar(255)")
     @Enumerated(EnumType.STRING)
-    private OrderStatus status;
+    private EatInOrderStatus status;
 
     @Column(name = "order_date_time", nullable = false)
     private LocalDateTime orderDateTime;
 
     @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
     @JoinColumn(
-        name = "order_id",
+        name = "eat_in_order_id",
         nullable = false,
         columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_order_line_item_to_orders")
+        foreignKey = @ForeignKey(name = "fk_order_line_item_to_eat_in_orders")
     )
     private List<OrderLineItem> orderLineItems;
-
-    @Column(name = "delivery_address")
-    private String deliveryAddress;
 
     @ManyToOne
     @JoinColumn(
         name = "order_table_id",
         columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_orders_to_order_table")
+        foreignKey = @ForeignKey(name = "fk_eat_in_orders_to_order_table")
     )
     private OrderTable orderTable;
 
     @Transient
     private UUID orderTableId;
 
-    public Order() {
+    public EatInOrder() {
     }
 
     public UUID getId() {
@@ -72,19 +64,11 @@ public class Order {
         this.id = id;
     }
 
-    public OrderType getType() {
-        return type;
-    }
-
-    public void setType(final OrderType type) {
-        this.type = type;
-    }
-
-    public OrderStatus getStatus() {
+    public EatInOrderStatus getStatus() {
         return status;
     }
 
-    public void setStatus(final OrderStatus status) {
+    public void setStatus(final EatInOrderStatus status) {
         this.status = status;
     }
 
@@ -102,14 +86,6 @@ public class Order {
 
     public void setOrderLineItems(final List<OrderLineItem> orderLineItems) {
         this.orderLineItems = orderLineItems;
-    }
-
-    public String getDeliveryAddress() {
-        return deliveryAddress;
-    }
-
-    public void setDeliveryAddress(final String deliveryAddress) {
-        this.deliveryAddress = deliveryAddress;
     }
 
     public OrderTable getOrderTable() {
