+++ src/main/java/kitchenpos/eatinorders/todo/domain/orders/EatInOrderRepository.java
@@ -1,0 +1,18 @@
+package kitchenpos.eatinorders.todo.domain.orders;
+
+import kitchenpos.eatinorders.todo.domain.ordertables.OrderTable;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface EatInOrderRepository {
+    EatInOrder save(EatInOrder order);
+
+    Optional<EatInOrder> findById(UUID id);
+
+    List<EatInOrder> findAll();
+
+    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, EatInOrderStatus status);
+}
+
