+++ src/main/java/kitchenpos/eatinorders/todo/domain/orders/EatInOrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.eatinorders.todo.domain.orders;
+
+public enum EatInOrderStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
